+++ src/execution/executor_update.h
@@ -250,10 +250,10 @@ public:
                 if (set_clause.op == UpdateOp::ASSIGNMENT) {
                     if (col_meta.type == TYPE_INT && rhs_col_meta.type == TYPE_FLOAT) {
                         *reinterpret_cast<int*>(data) =
-                            static_cast<int>(*reinterpret_cast<float*>(old_rec.data + rhs_col_meta.offset));
+                            static_cast<int>(*reinterpret_cast<double*>(old_rec.data + rhs_col_meta.offset));
                     } else if (col_meta.type == TYPE_FLOAT && rhs_col_meta.type == TYPE_INT) {
-                        *reinterpret_cast<float*>(data) =
-                            static_cast<float>(*reinterpret_cast<int*>(old_rec.data + rhs_col_meta.offset));
+                        *reinterpret_cast<double*>(data) =
+                            static_cast<double>(*reinterpret_cast<int*>(old_rec.data + rhs_col_meta.offset));
                     } else if (col_meta.type == TYPE_STRING || col_meta.type == TYPE_DATETIME) {
                         if (rhs_col_meta.type != TYPE_STRING && rhs_col_meta.type != TYPE_DATETIME) {
                             throw IncompatibleTypeError(coltype2str(col_meta.type), coltype2str(rhs_col_meta.type));
@@ -269,7 +269,8 @@ public:
                         if (rhs_col_meta.type != TYPE_FLOAT) {
                             throw IncompatibleTypeError(coltype2str(col_meta.type), coltype2str(rhs_col_meta.type));
                         }
-                        *reinterpret_cast<float*>(data) = *reinterpret_cast<float*>(old_rec.data + rhs_col_meta.offset);
+                        *reinterpret_cast<double*>(data) =
+                            *reinterpret_cast<double*>(old_rec.data + rhs_col_meta.offset);
                     }
                     continue;
                 }
@@ -279,12 +280,12 @@ public:
                     throw IncompatibleTypeError(coltype2str(rhs_col_meta.type), coltype2str(set_clause.rhs.type));
                 }
 
-                float base = rhs_col_meta.type == TYPE_INT
-                                 ? static_cast<float>(*reinterpret_cast<int*>(old_rec.data + rhs_col_meta.offset))
-                                 : *reinterpret_cast<float*>(old_rec.data + rhs_col_meta.offset);
-                float delta = set_clause.rhs.type == TYPE_INT ? static_cast<float>(set_clause.rhs.int_val)
-                                                              : set_clause.rhs.float_val;
-                float result = base;
+                double base = rhs_col_meta.type == TYPE_INT
+                                  ? static_cast<double>(*reinterpret_cast<int*>(old_rec.data + rhs_col_meta.offset))
+                                  : *reinterpret_cast<double*>(old_rec.data + rhs_col_meta.offset);
+                double delta = set_clause.rhs.type == TYPE_INT ? static_cast<double>(set_clause.rhs.int_val)
+                                                               : set_clause.rhs.float_val;
+                double result = base;
                 switch (set_clause.op) {
                 case UpdateOp::SELF_ADD:
                     result = base + delta;
@@ -311,7 +312,7 @@ public:
                     *reinterpret_cast<int*>(data) = static_cast<int>(result);
                     break;
                 case TYPE_FLOAT:
-                    *reinterpret_cast<float*>(data) = result;
+                    *reinterpret_cast<double*>(data) = result;
                     break;
                 case TYPE_STRING:
                 case TYPE_DATETIME:
@@ -333,9 +334,9 @@ public:
             }
             case TYPE_FLOAT: {
                 if (set_clause.rhs.type == TYPE_FLOAT) {
-                    *(float*)data = set_clause.rhs.float_val;
+                    *(double*)data = set_clause.rhs.float_val;
                 } else {
-                    *(float*)data = (float)set_clause.rhs.int_val;
+                    *(double*)data = static_cast<double>(set_clause.rhs.int_val);
                 }
                 break;
             }
